Integrate mentions and hovering toolbar features & Add answer page
